test: extract helper functions
